feat(platform): move SCION Microfrontend Platform to a separate Git repository
The development of the SCION Microfrontend Platform has started in the Git repository of the SCION Workbench. However, since the SCION Microfrontend Platform is an entirely independent project with no dependency on SCION Workbench, we have moved the SCION Microfrontend Platform to a separate Git repository. It further facilitates the maintenance of a distinct changelog and to follow the semantic versioning scheme for its releases.