chore: update project workspace to Angular 10
The project workspace of the SCION Microfrontend Platform depends on the Angular framework because:
- the internal test app is an Angular application
- the project is packaged using Angular ng-packagr

At runtime, however, the platform has no dependence on the Angular framework.

closes #19

BREAKING CHANGE: esm5 and fesm5 format is no longer distributed in `@scion/microfrontend-platform`’s NPM package

We no longer include the distributions for `esm5` and `fesm5` in the `@scion/microfrontend-platform`’s NPM package. Only the formats for `esm2015`, `fesm2015`, and UMD are distributed. Consequently, the module field in package.json now points to the `fesm2015` distribution.

To migrate:
- If requiring `esm5` or `fesm5`, you will need to downlevel to ES5 yourself. If using Angular, the Angular CLI will automatically downlevel the code to ES5 if differential loading is enabled in the Angular project, so no action is required from Angular CLI users.