ci: prevent Protractor from automatically updating the chromedriver to be compatible with Chromium
In `angular.json` we instruct Protractor not to automatically update the chromedriver to its latest version at startup.
Instead, we instruct the Selenium webdriver-manager to download the exact version of the chromedriver that is compatible with the Chromium provided by Puppeteer.

For more information, see https://github.com/angular/protractor/issues/4695#issuecomment-613718330.

closes #26