fix(platform): ensure trailing slash in application base URLs
Outlet router allows passing a base for computing the absolute navigation URL for a relative path. To do so, it must differentiate between a hash-based or push-state base, requiring the base to end with a slash. For that, we prefix application base URLs with a trailing slash, allowing us to solely specify the hash symbol (`#`) as the base for hash-based enabled applications.

We have also simplified the resolution of relative navigational symbols since JavaScript supported this natively. However, it turned out that the JavaScript URL object has a problem if the base does not end with a trailing slash, cutting off the last base segment. For this reason, we added a utility to construct URL objects.