chore: support path overrides for specific secondary entry points
This change affects the development of the SCION Microfrontend Platform library. Path overrides in `tsconfig` are configured to load dependent modules from the `node_modules` folder unless a path override for a specific primary or secondary entry point is commented out. The most specific paths win, or the  latter with equal specificity.