feat(platform): open shadow DOM of the router outlet
E2E tests require access to the iframe to change the test execution context. So far, we have exposed the iframe object via a member of the `SciRouterOutlet` element. However, it turned out that this is not too obvious. Instead, we were asked to open the Shadow DOM to select the iframe across the Shadow Root boundary using a CSS selector.

Opening the Shadow DOM does not affect microfrontend integration. Perhaps it was too academic to close the Shadow DOM since the iframe already gives full encapsulation. Finally, the router outlet only contains that iframe, so no need for additional encapsulation.