refactor(platform): remove gateway iframe in client-broker communication
The rationale for this extra gateway-iframe was to create a separate namespace for the platform and the host application in terms of communication and capabilities, and to allow multiple connections from a microfrontend.
But, in libraries such as the SCION Workbench, we cannot hook into the platform's internal namespace. Further, multiple connections from a microfrontend are a very hypothetical construct. Therefore, we decided to remove this extra iframe, which halves the number of iframes, eliminates a communication hop, and drastically reduces the complexity in the client-broker communication.

closes #14

BREAKING CHANGE: Removing the gateway communication iframe introduced a breaking change in the host/client communication protocol.

You need to upgrade the version of SCION Microfrontend Platform in host and client applications at the same time.
The breaking change refers only to the communication protocol, the API of the SCION Microfrontend Platform has not changed.

To migrate, upgrade to the newest version of `@scion/microfrontend-platform` in the host and client applications.