deps(platform): update project workspace to Angular 13 and migrate @scion/microfrontend-platform to RxJS 7.5
For more information, refer to
https://github.com/angular/angular/blob/master/CHANGELOG.md#1312-2022-01-12
or https://blog.angular.io/angular-v13-is-now-available-cce66f7bc296.

closes #108

BREAKING CHANGE: Upgrading @scion/microfrontend-platform to RxJS 7.x
introduced a breaking change.

The breaking change only refers to updating RxJS to version 7.5. API and
communication protocol have not changed or are backward compatible. Host
and clients can be updated independently to the new version.

To migrate:
- migrate your application to RxJS 7.5; for detailed migration
  instructions, refer to https://rxjs.dev/6-to-7-change-summary.