fix(platform): connect to the host periodically when starting the platform
When connecting to the host and if the client does not receive connect acknowledgement in time, the client will resend the connect request.

A single connect request may not be sufficient if, for example, the microfrontends are to be integrated into a rich client. For example, an integrator may want to bridge messages to a remote host. If the integrator cannot hook into the message bus in time, the client's connection request may be lost. Therefore, the gateway initiates the connection request at regular intervals until it receives an acknowledgement.