fix(platform): allow configuring the message origin for an application
The message origin defines where message from an application must originate from. Messages of a different origin will be rejected. If not set, the origin is derived from the manifest URL or the base URL as specified in the manifest file. Setting a different origin may be necessary if, for example, integrating the microfrontends into a rich client, allowing an integrator to bridge messages to a remote host.