fix(platform): remove `semver` dependency to avoid bundle optimization bailouts
We no longer depend on the NPM package `semver` as not yet been published as an ECMAScript module (ESM).
It is recommended to avoid dependencies on CommonJS modules as bundlers and minifiers may not be able to optimize the application code, resulting in larger bundle size.

For example, the Angular compiler would generate the following warning when using `semver`:
`Warning: ... depends on 'semver'. CommonJS or AMD dependencies can cause optimization bailouts.`

closes #141