feat(platform/router): facilitate intent-based navigation
The router now facilitates intent-based navigation as an alternative to direct navigation to a URL. We recommend using intent-based navigation, especially for cross-application navigation, since the navigation flows are explicit, i.e., declared in the manifest, and to keep the microfrontend URLs an implementation detail of the micro apps that provide the microfrontends.

Note that the idea of intent-based navigation is not new and can be implemented using the intent client and activator.

## Usage:

```ts
Beans.get(OutletRouter).navigate({entity: 'product'}, {params: {id: '123'}});
```

## Microfrontend Capability:
```json
 {
    "type": "microfrontend",
    "qualifier": {
      "entity": "product"
    },
    "params": [
      {"name": "id", "required": true}
    ],
    "properties": {
      "path": "product/:id",
    }
  }
```

closes #154