fix(platform/router): do not patch URL if protocol is "blob"
Due to a bug in the Chrome browser, the router adds a stable query parameter to the URL so that content from the same application can be loaded in nested iframes. However, the patch must not be applied to blob URLs to avoid corrupting them. See `RouterOutletUrlAssigner` or https://stackoverflow.com/q/36985731 for more information.