perf(platform/client): resend connect request only if loaded into the topmost window
Previously, repeated connect requests were also sent for clients that were included in a window hierarchy. This is not necessary and can lead to performance degradation, in particular when multiple clients want to connect to the host at the same time.

Sending repeated connect requests is required only if the client is loaded into the topmost window. In this case, the client may be integrated into a rich client, with the host running in a different browser window (remote host). The rich client then bridges messages between the windows of the client and the remote host. Since the rich client may not be able to bridge messages right away when the client loads, the client repeatedly sends a connect request until acknowledged by the remote host.