feat(platform/messaging): support sending retained intents
Similar to a retained message, a retained intent helps newly subscribed clients receive the last intent published for a capability. The platform supports the sending of retained intents and requests.

closes #142

BREAKING CHANGE: The `IntentClient` now uses the same `options` object as the `MessageClient` to control how to send a message or request. The content of the `options` object has not changed, only its name, i.e., migration is only required if assigning it to a typed variable or decorating the `IntentClient`.

Note that the communication protocol between host and client has NOT changed, i.e., host and clients can be migrated independently.

To migrate:
 - options for sending an intent was changed from `IntentOptions` to `PublishOptions` (`IntentClient#publish`)
 - options for requesting data was changed from `IntentOptions` to `RequestOptions` (`IntentClient#request`)