fix(platform/messaging): stop platform in `beforeunload` to avoid posting messages to disposed window
When posting a message to a disposed window, the browser logs "Failed to execute 'postMessage' on 'DOMWindow'" error.
To avoid posting messages to disposed windows, we now disconnect clients already in `beforeunload` instead of `unload`.
However, if `beforeunload` is not triggered, e.g., when an iframe is removed, we fall back to `unload`.

closes #168