fix(platform): do not unregister clients after resuming the computer from hibernation
closes #178

BREAKING CHANGE: The host now performs liveness probes to detect and
remove stale clients, instead of relying on the heartbeats emitted by
the clients of previous versions.

The breaking change only refers to the host. The communication protocol
between host and client has NOT changed. You can independently update
host and clients to the new version.

To migrate setting of a custom probe configuration in the host, specify
the `liveness` instead of `heartbeatInterval` property, as follows:

```ts
MicrofrontendPlatform.startHost({
  liveness: {interval: 60, timeout: 10},
  // omitted rest of the config
});
```