fix(platform/client): remove startup progress monitor from client
BREAKING CHANGE: removing startup progress monitor from client introduced a breaking change in the client.

To reduce the size of the client bundle, the API for monitoring the startup progress in the client has been removed. This API is useful in the host, but not in the client, because the client only connects to the host and does not perform any long-running operation.