perf(platform): reduce library bundle size in client application
closes #174

BREAKING CHANGE: reducing bundle size in client application introduced a breaking change for host and client applications. The communication protocol between host and client has not changed, i.e., host and clients can be updated independently to the new version.

To enable tree-shaking of the SCION Microfrontend Platform, the platform was split into three separate entry points:
- `MicrofrontendPlatformHost` to configure and start the platform in the host
- `MicrofrontendPlatformClient` to connect to the platform from a microfrontend
- `MicrofrontendPlatform` to react to platform lifecycle events and stop the platform

The size of the library bundled in a client application could be reduced by more than 50%, from 120 KB to 40 KB.

To migrate the host application:
  - start the platform via `MicrofrontendPlatformHost.start` instead of `MicrofrontendPlatform.startHost`
  - monitor startup progress via `MicrofrontendPlatformHost.startupProgress$` instead of `MicrofrontendPlatform.startupProgress$`

To migrate the client application:
  - connect to the host via `MicrofrontendPlatformClient.connect` method instead of `MicrofrontendPlatform.connectToHost`
  - test if connected to the host via `MicrofrontendPlatformClient.isConnected` method instead of `MicrofrontendPlatform.isConnectedToHost`