deps(platform): update project workspace to Angular 15
Previously, to test microfrontends in unit tests, we used Webpack's `require.context` function to transpile and serve script files that could be loaded into iframes during test execution. As of Angular 15, support for `require.context` has been discontinued, and custom Webpack configurations are not guaranteed to work in the future.

For this reason, we have migrated as follows:
  - We use Karma testrunner to serve and watch script files.
  - We use `karma-esbuild` plugin to transpile script files into js files using `esbuild`.
  - Transpiled script files are served by Karma testrunner.

For more information, refer to
- https://v15.angular.io/guide/update-to-latest-version
- http://karma-runner.github.io/6.4/config/files.html
- https://www.npmjs.com/package/karma-esbuild
- https://esbuild.github.io/api/

closes #169